% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{initialize_logger}
\alias{initialize_logger}
\title{Learnr Logging Handlers}
\usage{
initialize_logger()
}
\description{
Functions to register handlers that log session and exercise events to a
Cloudflare Worker endpoint.
}
